requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
huggingface-hub
pypdf
python-dotenv
langchain-google-genai
